Refactor thread copy logic into useCopyThread hook
Moved the thread copying functionality from MainApp into a new reusable useCopyThread hook. This improves code organization and reusability by encapsulating the clipboard and error handling logic.
